move the code to Ranking according to property name
