+++ container-search/abi-spec.json
@@ -5775,6 +5775,8 @@
       "public com.yahoo.search.query.ranking.Matching getMatching()",
       "public com.yahoo.search.query.ranking.SoftTimeout getSoftTimeout()",
       "public com.yahoo.search.query.ranking.Significance getSignificance()",
+      "public java.lang.Integer getElementGapForField(java.lang.String)",
+      "public void setElementGapForField(java.lang.String, java.lang.Object)",
       "public com.yahoo.search.query.Sorting getSorting()",
       "public void setSorting(com.yahoo.search.query.Sorting)",
       "public void setSorting(java.lang.String)",
@@ -5807,7 +5809,8 @@
       "public static final java.lang.String SOFTTIMEOUT",
       "public static final java.lang.String MATCHING",
       "public static final java.lang.String FEATURES",
-      "public static final java.lang.String PROPERTIES"
+      "public static final java.lang.String PROPERTIES",
+      "public static final java.lang.String ELEMENT_GAP"
     ]
   },
   "com.yahoo.search.query.Select" : {
@@ -7405,8 +7408,6 @@
       "public void setExplorationSlack(double)",
       "public void setTargetHitsMaxAdjustmentFactor(double)",
       "public void setFilterThreshold(double)",
-      "public java.lang.Integer getElementGapForField(java.lang.String)",
-      "public void setElementGapForField(java.lang.String, java.lang.Object, com.yahoo.search.Query)",
       "public void prepare(com.yahoo.search.query.ranking.RankProperties)",
       "public com.yahoo.search.query.ranking.Matching clone()",
       "public boolean equals(java.lang.Object)",
@@ -7426,7 +7427,6 @@
       "public static final java.lang.String TARGET_HITS_MAX_ADJUSTMENT_FACTOR",
       "public static final java.lang.String FILTER_THRESHOLD",
       "public static final java.lang.String WEAKAND",
-      "public static final java.lang.String ELEMENT_GAP",
       "public java.lang.Double termwiseLimit"
     ]
   },
